requirements.txt
cffi==1.15.0
cryptography==36.0.2
noiseprotocol==0.3.1
pkg-resources==0.0.0
pycparser==2.21
semantic-version==2.9.0
setuptools-rust==1.1.2
typing-extensions==4.1.1
